Moving code from global convert.go to volumes.go, ports.go
